payment

+++ order/src/main/java/order_service/order/services/PricingService.java
@@ -1,8 +1,9 @@
 package order_service.order.services;
 
 import order_service.order.model.OrderItem;
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface PricingService {
-    double calculateTotalPrice(List<OrderItem> items);
+    BigDecimal calculateTotalPrice(List<OrderItem> items);
 }
